Reorganización del proyecto en diferentes módulos Agrego instrucciones al readme para arrancar el proyecto spring
