+++ README.md
@@ -1,2 +1,20 @@
 # TACS-2019-GRUPO3
 Repositorio del TP cuatrimestral de TACS UTN 1C-2019
+
+### Proyecto Spring
+
+#### Build
+
+Desde el root del proyecto
+```
+    cd grupo3/grupo3-spring-controller
+    mvn compile
+```
+
+#### Run
+
+Desde el root del proyecto
+```
+    cd grupo3/grupo3-spring-controller
+    mvn spring-boot:run
+```
