+++ grupo3/.gitignore
@@ -2,6 +2,11 @@ HELP.md
 /target/
 !.mvn/wrapper/maven-wrapper.jar
 
+### Intellij ###
+
+.iml
+.idea/
+
 ### STS ###
 .apt_generated
 .classpath
