+++ grupo3/grupo3-spark-controller/src/main/java/utn/tacs/grupo3/spark/Application.java
@@ -1,0 +1,17 @@
+package utn.tacs.grupo3.spark;
+
+import spark.Spark;
+import spark.debug.DebugScreen;
+import utn.tacs.grupo3.spark.router.Router;
+
+public class Application {
+	
+	public static void main(String[] args) {
+		Spark.port(8080);
+		DebugScreen.enableDebugScreen();
+		Router.configure();
+
+	}
+
+}
+
