+++ grupo3/grupo3-spring-controller/src/main/java/utn/tacs/grupo3/spring/Grupo3Application.java
@@ -1,0 +1,15 @@
+package utn.tacs.grupo3.spring;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan(basePackages = {"utn.tacs.grupo3"})
+public class Grupo3Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Grupo3Application.class, args);
+	}
+	
+}
