+++ grupo3/grupo3-spring-controller/src/test/java/utn/tacs/grupo3/spring/Grupo3ApplicationTests.java
@@ -1,0 +1,16 @@
+package utn.tacs.grupo3.spring;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class Grupo3ApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
