+++ grupo3/pom.xml
@@ -2,63 +2,29 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<parent>
+    <packaging>pom</packaging>
+    <modules>
+        <module>grupo3-spark-controller</module>
+        <module>grupo3-spring-controller</module>
+        <module>grupo3-model</module>
+        <module>grupo3-repository</module>
+        <module>grupo3-foursquare-integration</module>
+    </modules>
+    <parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.1.3.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>utn.tacs</groupId>
-	<artifactId>grupo3</artifactId>
+	<artifactId>utn.tacs.grupo3</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>grupo3</name>
-	<description>Demo project for Spring Boot</description>
+	<name>utn.tacs.grupo3</name>
+	<description>TP del grupo 3 de la materia TACS UTN FRBA 2019 1C</description>
 
 	<properties>
 		<java.version>1.8</java.version>
 		<spark.version>2.8.0</spark.version>
 	</properties>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-web</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-		
-		<!-- #################### SPARK #################### -->
-		<dependency>
-			<groupId>com.sparkjava</groupId>
-			<artifactId>spark-core</artifactId>
-			<version>${spark.version}</version>
-		</dependency>
-		
-		<dependency>
-			<groupId>com.sparkjava</groupId>
-			<artifactId>spark-debug-tools</artifactId>
-			<version>0.5</version>
-		</dependency>
-
-		<dependency>
-			<groupId>com.google.code.gson</groupId>
-			<artifactId>gson</artifactId>
-			<version>2.8.5</version>
-		</dependency>
-		<!-- #################### SPARK #################### -->
-	</dependencies>
-
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.springframework.boot</groupId>
-				<artifactId>spring-boot-maven-plugin</artifactId>
-			</plugin>
-		</plugins>
-	</build>
 
 </project>
