--- grupo3/src/main/java/server/Router.java
@@ -1,19 +1,0 @@
-package server;
-
-import com.google.gson.Gson;
-import controller.AdminController;
-import spark.Spark;
-
-public class Router {
-	public static void configure() {
-		Spark.staticFiles.location("/public");
-		Gson gson = new Gson();
-
-		AdminController adminController = new AdminController();
-		Spark.get("/usuarios/:id", adminController::getInformacionDelUsuario, gson::toJson);
-		Spark.get("/lugares_en_comun/usuarios", adminController::getLugaresEnComun, gson::toJson);
-		Spark.get("/lugares/:id/usuarios_interesados", adminController::getCantidadDeUsuariosInteresadosEnUnLugar, gson::toJson);
-		Spark.get("/lugares/registrados_hasta_la_fecha", adminController::getRegistrosTotalesEnUnaFecha, gson::toJson);
-	}
-
-}
