--- grupo3/src/main/java/server/Server.java
@@ -1,16 +1,0 @@
-package server;
-
-import spark.Spark;
-import spark.debug.DebugScreen;
-
-public class Server {
-	
-	public static void main(String[] args) {
-		Spark.port(8080);
-		DebugScreen.enableDebugScreen();
-		Router.configure();
-
-	}
-
-}
-
