--- grupo3/src/test/java/utn/tacs/grupo3/Grupo3ApplicationTests.java
@@ -1,16 +1,0 @@
-package utn.tacs.grupo3;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class Grupo3ApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
